Resuce domain error
NoMorePhaseSpec
CardOnPhase

+++ app/commands/remove_phase_spec_command.rb
@@ -3,6 +3,9 @@ class RemovePhaseSpecCommand
 
   attr_accessor :project_id_str, :phase_name
 
+  validates :project_id_str, presence: true
+  validates :phase_name, presence: true
+
   def project_id
     Project::ProjectId.new(project_id_str)
   end
@@ -12,6 +15,15 @@ class RemovePhaseSpecCommand
   end
 
   def execute(service)
+    return false unless valid?
     service.remove_phase_spec(project_id, phase)
+  rescue Project::NoMorePhaseSpec
+    errors.add(:base, 'フェーズが1つしかないため削除できません。')
+    false
+  rescue Project::CardOnPhase
+    errors.add(:base, 'フェーズにカードがあるため削除できません。')
+    false
+  else
+    true
   end
 end
